Allow fractional opponent scoring ranks

+++ src/ncaa_rankings/models/base.py
@@ -85,7 +85,7 @@ class RankingModel(Protocol):
     def score_game(
         self,
         game: TeamGame,
-        opponent_rank: int | None,
+        opponent_rank: float | None,
         team_count: int,
         *,
         age_weeks: float = 0.0,
